Removed disgusting and soon to be old level loading code from map to map_from_file
